Change the error handling model.
The core error handler is now a chip which accepts
an error message, prints it, and terminates the process.

We use a wrapper procedure which throws a procedures
which writes a message, and pass that to functions and procedures
as a parameter.

Calling that procedure binds the error message to the inner
procedures which is then thrown. When it is caught by the scheduler,
it becomes the current continuation, and writes the message to
the error handling chip which prints it, and then terminates
the processes.

The procedure can be invoked by any procedure or function
anywhere, however it cannot be used in expressions without
yet another wrapper which forces the return type to agree
with other match branches or the return type.

The return type "any" should correctly unify, however the
application doesn't seem to correctly coerce it to the
required type. Either that or another hack is getting
in the way: Felix is converting a function returning
any into a procedure call, which returns type 0, which
of course won't work.

It would be nice if this were working better but I won't
try to fix it now because it could break something else,
and it's easy to work around.