Fix BREAK. Hack collatz.mid to put dummy LABEL in.
There has to be a LABEL soon after a GOTO or BREAK
to force control to run down the if/then/else chain
to the end of the SWITCH loop. The compiler should put
these in automatically but doesn't at the moment.

The BREAK instuction forgot to dup the PC so got removed
from the stack when setting the PC variable. An extra dup
fixed that. The modified collatz.mid now compiles and runs.