The code is modified to use the new error handling but fails.
This branch is intended to act as a test case for upgrades to
the Felix compiler. A new feature was previously partially supported,
which allowed a procedure to return a type "any". This return type
represents a functional unit which does not return control and
so it's invocation can be used to escape the current context.

The key purpose of the typing is to allow the escape to be expressed
in *either* a procedural or functional context with the same construction;
in other words `fail "error"` can be used as an expression OR as
a statement.

This works in when the invocation is direct, however the back end
thinks the function is a procedure, and, even inside a function,
is trying to call it as one, passing it the current continuation
as it's return address when (a) the current continuatino of a function
in Felix is not accessible because it's on the machine stack, AND,
(b) the function doesn't return anyhow. So there seem to be TWO
distinct problems.

These are NOT bugs: the modifications are work in progress.
There bascially only two things you can do to escape:
(a) call System::exit to terminate the process immediately
(b) C++ throw a continuation

Throwing a continuation uses C++ to unwind the machine
stack up to the nearest scheduler which traps exceptions.
The top level scheduler traps exceptions. Normal procedure
calls which are not optimised to C procedures are called
using a compiler emitted micro-scheduler which does not
trap exceptions. A special instruction, call_with_trap,
can be used to insert the exception handling.

Because the function returning "any" cannot return,
it should probably be implemented by a Felix function call
rather than a procedure call, since this works in ANY
context.

So the current issue appears to be well understood an easy
to fix ... famous last words!